requirements.txt
django_rest_swagger==2.2.0
Django==3.2
drf_spectacular==0.20.1
djangorestframework==3.12.4
psycopg2==2.9.1